[ExampleTicket#1]: Create Nested Stacks for arch tiers
In this commit, I've created nested Stacks for each different tier within the architecture (application, visibility, management). This is to reduce the blast radius on anything that might need rolling back. Text from CDK docstrings:
When you apply template changes to update a top-level stack, CloudFormation updates the top-level stack and initiates an update to its nested stacks. CloudFormation updates the resources of modified nested stacks, but does not update the resources of unmodified nested stacks.

Furthermore, this stack will not be treated as an independent deployment artifact (won't be listed in "cdk list" or deployable through "cdk deploy"), but rather only synthesized as a template and uploaded as an asset to S3.

Reported-by: seanmca
Signed-off-by: Sean McAuliffe smcauliffe240@gmail.com